+++ CHANGELOG.md
@@ -11,7 +11,9 @@ The official release will have a defined and more stable API. If you are already
     * `lambdacd.presentation.pipeline-state/overall-build-status`
     * `lambdacd.presentation.pipeline-state/latest-most-recent-update` 
     * `lambdacd.presentation.pipeline-state/earliest-first-update` 
-    * `lambdacd.presentation.pipeline-state/build-duration` 
+    * `lambdacd.presentation.pipeline-state/build-duration`
+    * Renames namespace `lambdacd.internal.step-id` to `lambdacd.step-id` to officially make it public. 
+      In case someone was using the internal namespace, it is still there but is considered DEPRECATED and will be removed in subsequent releases. 
   * UI: Link LambdaCD header to "/" to link back to overview in cases with multiple pipelines (#82)
 * Bugs: 
   * `with-workspace` now creates temporary directories in the home-dir (#79)